Fixed configure to not require lex library
Fixed configure script to not fail when flex library (-lfl or -ll) is
not installed, since we use %option noyywrap and do not need it.

* configure.ac: Replaced AC_PROG_LEX with AC_CHECK_PROGS to find flex
  binary without checking for lex library
* configure.ac: Set LEXLIB empty since we use %option noyywrap
* configure.ac: Set LEX_OUTPUT_ROOT for compatibility

AC_PROG_LEX was causing configure to fail with "required lex library
not found; giving up on flex" even though flex 2.6.4 was installed.
This happened because the system did not have libfl or libl installed,
which is not needed when using %option noyywrap in lexer files.

The new approach simply checks for the flex or lex binary without any
library dependency checking.